Add an `enumerate` combinator
Returns the index into the original sequence along with each item.
Works like `Iterator::enumerate()`, but of course with threaded jobs it
won't be visited in any guaranteed order.  It has to assume that the
size of each split is known exactly though, so this wouldn't work after
a filtering function, for instance.