impl ParallelIterator for Range<T> (for some T)
A par_iter/range.rs was already started, but not actually pulled in.

Range<T> is annoying to deal with generically, since its Iterator
depends on unstable Step, and its ExactSizeIterator is only for select
integral types (all but i64/u64).

Use a macro to implement just those Range<T> with ExactSizeIterator.