Add an experimental parallel fold combinator
Fold is very similar in use to iter.map().reduce() but somewhat more
"fused together". The iterator is divided up and the smaller pieces are
processed sequentially to produce results (in this case, the combined
diff due to a list of bodies). Those results are then merged using the
reduce operation.

This gives a mild speedup, but perhaps not enough to justify its
existence. May need more tweaking.