Refactor VecIter unsafety so it doesn't need Drop
The `consumed` flag really only served to let `with_producer` avoid
unsafe code itself.  But that made an implicit safe-unsafe contract with
the drop handler, which isn't good encapsulation.

Now all of the `VecIter` unsafety is contained in `with_producer`, and
it no longer needs a `Drop` implementation at all.