add (unstable) **scoped threads** API
Whereas the `join` API also spawns (and joins) two tasks, the scoped
thread API allows you to spawn off an unlimited number of tasks. These
tasks will be joined before the scope completes. Unlike the `join` API,
these tasks are allocated on the heap. This is more expensive, but can
be useful if you do not know the number of tasks in advance.