RFC: adaptive "thief splitting" algorithm for par_iter
This introduces an adaptive algorithm for splitting par_iter jobs, used
only as a default case when no weights are specified.  Initially, it
will split into enough jobs to fill every thread.  Then whenever a job
is stolen, that job will again be split enough for every thread.

This is roughly based on @Amanieu's description in the users forum of
the algorithm used by Async++ and TBB.