a parallel TSP solver based on the scope construct
This is not intended as a truly competitive implementation, at least not
yet =), but more as a stress test and general benchmark. It works by
exhaustively exploring the space of possible tours in parallel. At each
point, it retains a (not very clever) heuristic of a lower bound on the
cost to complete the tour. Behind the scenes, a priority queue is used
so that each parallel thread is exploring the most profitable
possibilities. After a certain point, we fall back to sequential
execution.

The current implementation uses the `scope` construct. There are various
enhancements I would like to make:

- perhaps other lower bound heuristics, such as leveraging the triangle inequality
- replace the mutex with a flat combining lock, once we offer those
- make a "pure sequential" variant to compare against, although simply
  setting a sequential threshold of N may be "good enough"
- replace the manual sequential threshold with a dynamic variant,
  not sure yet what this would look like
  - I imagine something like pushing a minimum amount of work
  - and then some kind of quick check on the length of the local queue
    to decide whether to push more