strengthen/justify memory orderings
After analyzing the scenario more deeply, I found that, for correctness,
both the latches and sleep code require SeqCst memory
orderings. However, we can permit weaker orderings in the deque
code (which is out of our control anyhow). The performance impact of
these changes was found to be negligible (but using stronger orderings
for the deque code did certainly have an impact).