Remove the fake specialization on chain collect
`ChainIter::drive_unindexed` can't deal with the fake specialization
used by `CollectConsumer`, as it has to directly call `split_off`.  So
we must not report an `opt_len`, and thus iterators with any chaining
can't take advantage of that optimization.

However, it would be a regression if we couldn't even use `collect_into`
with chains, so we need to let that call a direct indexed `drive` again.
So refactor the entry points with `collect_into` calling `drive`, and
still `special_collect_into` calling `drive_unindexed`, with all of
their common code in a new `Collect` wrapper.