Fork into `rayon-core` and a stable `rayon` facade
The root crate is now `rayon-core`, with all functionality, and under
`rayon-stable/` is the new `rayon` crate that only publicizes select
really-stable interfaces -- those we feel are nearly ready for 1.0.  For
the most part, this means the APIs for *using* rayon, while we can
continue tweaking the APIs for *implementing* rayon traits in the core
with proper pre-1.0 semver bumps.

(The directories are structured this way just to minimize the code churn.)

Most of the traits are now split, e.g. `ParallelIterator` and a new
`ParallelIteratorImpl`, so only the former is publicized in `rayon`.
Methods like `drive` and `with_producer`, anything that implementors
have to provide, are isolated in `*Impl`.