Require `Send` output in `map`, `cloned`, `filter_map`
All three of these methods lacked any constraints on their output types,
even though it has to be `Send` for the returned adaptor to itself be a
`ParallelIterator`.  This sometimes leads to error messages in usage
that give no indication that missing `Send` is the real problem.  If we
add that requirement up front, it should be a more helpful error.

Fixes #204.