Add a new `rayon::split` to create arbitrary parallel iterators
This takes two arguments, the data and a closure to split it, and
implements a `ParallelIterator`.  The data will be divided using the
usual `par_iter` splitting heuristics, then may be mapped, reduced, etc.
however is needed.

The `fibonacci` and `quicksort` demos have been updated with `split`
examples, where the former splits `n` into `n-2` and `n-1`, and the
latter performs a sorting partition before splitting its slice.