Fix the reduction order of flat_map
Usually in `flat_map`, the cost is infinite, so `FlatMapFolder` doesn't
expect to ever see more than one item.  However, `UnindexedProducer` is
only using the thief splitter, never weights, so it doesn't notice the
cost and may in fact have multiple items in this fold.

It did handle this possibility just in case, but the order of the reduce
arguments were swapped, with the new result before the previous.  This
is easily seen with collect, where items show up in the wrong order.

(Generally, we require that ops are associative, but not commutative.)