Refactor global registry intialzation, forbid double initialization
Remove `pub` from `global_registry`. `get_registry_with_config` renamed
to `init_global_registry`. Forbid double initialization of the global
registry, we allowed double initialization if the configurations were
equal and do nothing on the second call. However that's bogus because
`Configuration` does not implement `PartialEq` and we  would
incorrectly assume that if the thread_num is equal then the
configuration is equal.