Create unique parallel iterators for all collections
For all of the `std::collections`, create distinct parallel iterators
with the same path in rayon, e.g. `rayon::collections::hash_map::Iter`.
Most of these are still implemented by vectorizing, using a sequential
iterator to collect a vector, and then iterating *that* in parallel.
But now this detail is not actually exposed in the API, so we can feel
free to improve on this in the future without a breaking change.