Make an explicit delegate_indexed_iterator
We don't always want to expose the fact that our underlying iterator
is indexed, because that might not remain the case for a more native
implementation (i.e. not going through a vec intermediary).  So drop
that from the base `delegate_iterator!`, and use the new indexed version
where it makes sense for a type.