Abuse `links` to make sure there's only one rayon-core
One of the main points of having `rayon-core` is so we can stabilize it
~forever, and make sure there's only one instance running, as everyone
should agree on some 1.x.y.  But that only works for caret requirements;
one could still specify distinct versions with tilde, wildcard, or
inequality requirements.

There is a `links` manifest key for making sure only one crate links to
a particular native library.  We can abuse that to achieve a similar
effect for rayon-core without even linking anything.

    error: native library `rayon-core` is being linked to by more than
    one package, and can only be linked to by one package

As long as every published rayon-core specifies `links` this way, they
won't be allowed to coexist.