Remove redundant split_off_left() calls
In `FlatMapConsumer::split_at`, we can just pass it down to the base
consumer's `split_at`.  Avoiding the double `split_off_left()` here
might improve the indexing resolution of `find_first`/`last`, but is
otherwise just a small optimization.

In `ForEachConsumer::split_at`, we only need one `split_off_left()` to
get a copy of `self`, and then we can just use `self` for the second.