Remove bounded and exact iterators
This reduces our iterator traits to just `ParallelIterator` and
`IndexedParallelIterator`.  The bounded and exact iterators had some
theoretical use, but in practice the only distinction we use today is
whether something can be indexed.

The `upper_bound` method is totally removed.  The rest of the methods in
the removed traits are now merged into `IndexedParallelIterator`.

Fixes #229.
Closes #259.