Clarify the ordering of find_first etc.
It's not fully clear that "first" in `find_first` and `position_first`
is referring to the first in the sequential order of items, rather than
the temporally first match found anywhere like `find_any`.  This update
emphasizes that we mean sequentially first.

There shouldn't be the same issue in `find_last` and `position_last`, as
there's no reason to look for a "temporally last" match, but those docs
are also updated just to keep similarity to the "first" methods.