Add `ParallelIterator: Send` to join `Chain`
If the parallel iterators are themselves always `Send`, then we can use
`join` in implementing `Chain::drive` and `drive_unindexed`.  This
resolves the current sequence point between each side, as they can now
run in parallel if a thread is available to steal it.

The main fallout of that new requirement is that closures we accept must
now be `Send` too, in addition to the existing `Sync` requirement.  It's
not strictly necessary in all cases, like `for_each` that doesn't create
a parallel iterator itself, but it seems nicer to apply a consistent
requirement.

Fixes #342.