Introduce `join_notify` which tells whether jobs were stolen
This removes the need for tracking thread-id or hacky TLS variables in
the adaptive splitter.  Instead, with `join_notify` the parallel
iterator internals are directly informed whether a job is executing on
its original thread or stolen elsewhere.

Note that this is insta-stable since we want to use it ourselves.  We
could hide docs and/or mark it deprecated to approach stability a little
differently than the usual "unstable" feature.