Add `for_each_with`
I had considered this method when adding `map_with` and `fold_with`, but
at the time I decided it wasn't "primitive" enough.  It is literally
just `map_with(...).for_each(no-op)`.  But we already have a user asking
the best way to drive a void `map_with`, so it may be useful to provide.