Solve inter-pool deadlocks
`Registry::in_worker()` now calls a distinct `Registry::in_worker_cross`
when a `WorkerThread` *is* found for the current thread, but it's not
part of this `Registry`'s pool.  We use a new `TickleLatch` which will
awaken a pool when it's set.  Then the current thread can go process
other work from its own pool, possibly going to sleep when idle, and
still be notified when the other pool finishes the new work.