Add thorough tests for Producer::split_at
This takes one of the test cases from `intersperse`, which was very
helpful in poking corner cases, and generalizes it to test all
`Producer`s.

This found a semi-bug with `enumerate().rev()` in the process -- by
zipping with indexes up to `usize::MAX`, reverse iteration has to
iterate a loooong way back.  It now uses the exact length needed.