Break out try_reduce_with to avoid unwanted constraints
Even though we already know `Self::Item: Send`, this doesn't apply
constraints to its `Try::Ok` and `Error`.  Therefore, if we map to
`Result<Option<Ok>, Error>` at a high level, we have to require extra
`Send` bounds to make sure that map is `Send`, like:

     <Self::Item as Try>::Ok: Send,
     <Self::Item as Try>::Error: Send

We can avoid this by implementing a custom `Consumer`, where we only
hold it as a `Result` in the low level `Folder` regardless of `Send`.