Implement opt_len for ranges of large integers
If the length of a range of `i64`, `u64`, `i128`, or `u128` would
actually fit in `usize`, we can return that from `opt_len` to enable
some optimizations, namely indexed `collect` into vectors.  To drive
such ranges, we'll actually iterate over `(0..len)` instead, and then
map it to the larger type.