Add env vars for ROUNDS_UNTIL_SLEEPY and ROUNDS_UNTIL_ASLEEP
Running rayon in a rather tight loop can cause very high cpu and power
consumption that is not proportional to the amount of work being performed.
As documented in the sleep module, this is caused by the work stealing
algorithm and workers spinning in a hot loop while looking for work.

This commit adds two environment variables that serve as configuration
knobs and allows the user to adjust how aggressive the workers behave.

* `RAYON_ROUNDS_UNTIL_SLEEPY`
* `RAYON_ROUNDS_UNTIL_ASLEEP`