Trim the error codes in compile_fail/no_send_par_iter.rs
The compiler used to error E0277 on one line and E0599 on those that
follow, but recently it's only using E0277 on both lines. That's nice,
as it's the more informative error anyway in this case. We just need to
stop trying to match E0599 in the `compile_fail` test.