Fix TickleLatch reading stale data after set()
`TickleLatch` is used for synchronizing between multiple thread pools.
It sets an internal latch, and then tries to tickle the remote pool to
notice the completion. However, if the other pool is already awake and
sees the set flag, it may continue and drop the whole job, invalidating
further reads from the latch.

We can fix this by reading all fields from the `&self` latch _before_
calling the internal `set()`. Furthermore, I've updated this to take a
full `Arc<Registry>` handle, so we're also sure that the other pool
won't be destroyed too soon.

The new `tests/cross-pool.rs` would usually still pass before, but the
race could be increased by adding a manual `yield_now()` between the
`set()` and `tickle()`. This PR still passes under that same hack.

Fixes #739.