When collecting into a vector, track initialized values for correct drop on unwind
The collect folder produces a CollectResult<'c, T>,
This is a "proxy owner" of the slice of elements that the folder has
written to, and if unwinding happens under way, the elements will
be dropped.

Rewrite the Collect methods so that it is used like this:

    Collect::new(&mut vector, len).with_consumer(|consumer| ... );

Where the closure must run the collect consumer and return a
CollectResult of the promised length, to "prove" that the whole vector
has been written to.

The Collect can assert that the final CollectResult is covering the
target slice. The CollectResult borrows from the target slice, and the
lifetime parameter ensures that the collect caller can't subsitute or
fake the collect result.

The previous atomic used for counting writes is now removed, since
we trust the collected result instead. The producers under way
could misbehave, but in that case the CollectResult will not
end up correct.