find benchmark with blocks
we could win big by adding more operations.
it is super nice to have the sequential iterator on all reduced values
of blocks.
we could then use the dumb_find on each block and use the sequential
find to find the first value.

this would be the best algorithm.

sadly we can't have it due to rayon's types encapsulation.
the next best thing would be a
try_fold_by_exponential_blocks method
it would take a closure on Self producing the reduced values
and a closure for the sequential try_fold