Make sure `ParBridge` threads consume all items after `done`
We have the `done` flag to indicate when the bridged iterator is out of
items, but there may still be items left in the parallel deque as well!
Threads should check both, in that order, otherwise they may drop out of
the bridge party while there's still work to do.