Use in_place_scope to spawn skipped items
`Skip` still executes the skipped part of the iterator with a no-op
consumer in case there are side effects, as a point of compatibility
with sequential `Skip`. However, the `Producer` is not necessarily
`'static`, and `ProducerCallback::Output` doesn't require `Send`, so we
didn't have a way to run the no-op in parallel with the normal callback.

Now with `in_place_scope`, we can `spawn` the no-op consumer while we
run the callback directly. The scope will still wait for that no-op to
complete before returning, propagating panics if necessary.