Enforce THREADS_MAX and expose it with max_num_threads()
We have an inherent maximum on the number of threads we can support in a
single thread pool, based on the number of `THREADS_BITS` we are using
in the atomic sleep counters. However, we were not enforcing this, so a
larger pool could be created and end up wrapping the counters. In debug
builds this would fail an assertion, otherwise it may lead to errors in
the desired sleep and wakeup behavior.

We now assert this limit in `Sleep::new`, and also impose it as a soft
limit in `Registry::new`, automatically reducing to the maximum. A
top-level `pub fn max_num_threads()` returns the maximum for users.

At the same time, the value of `THREADS_BITS` is now adjusted for
different machine sizes. It was 10 bits everywhere, but 32-bit systems
now use 8 bits for max 255 threads (leaving more room for the JEC),
while 64-bit systems now use 16 bits for a maximum of 65,535 threads.