Be careful comparing `job_ref.execute_fn`
After #1025, clippy is denying [`fn_address_comparisons`][1], and this
is correct that we can't actually guarantee that the pointer we got from
`as_job_ref` will be the same that we produce in a later `PartialEq`.
It's less worrying that different functions could merge to the same
address, because our `StackJob` will still have different data pointers.

We can keep `Copy + Eq` off of `JobRef` with an opaque `id() -> impl Eq`
instead, so we'll be comparing the exact same captured `fn` pointer.

[1]: https://rust-lang.github.io/rust-clippy/master/index.html#fn_address_comparisons