Run tests w/ fallback on wasm32-unknown-unknown
With this change, CI can actually run tests against this target which should help catch any bugs early.

This is a bit invasive, because wasm32-unknown-unknown requires a custom `#[wasm_bindgen_test]` attributes as well as a custom runner instead of the standard `#[test]`, but there is no way around it, and, having those changes only at the top of test files, I believe it's reasonably isolated.