Use pseudo-specialization in all `ParallelExtend`
`Vec` originated this "specialization" based on `opt_len`, and the
recent `collect_vec_list` used the same technique. Now that is pushed
down a layer so that every `impl ParallelExtend` that uses a
`LinkedList<Vec<_>>` intermediate can also optimize the indexed case to
a single collected `Vec`, before extending the final collection.