Sync `par_sort*` with the standard library
Our parallel sort methods were already based on the standard library,
but they're out of date. Most of that implementation was moved to a
unified `core::slice::sort` module, with manual allocation functions
passed from the `alloc` crate and a simplified `Vec`.

This PR updates `rayon` to a unified `slice::sort` based on `core`'s,
but dropping the manual allocation stuff to just use the real `Vec`. The
rest of the changes are straightforward parallelization, like using `Fn`
instead of `FnMut`, and the extended `par_mergesort` additions are still
the same as we had before.