Fix incorrect overflow handling in binomial
In cases where `k` was 2 or 3, no guard was used, so this function could
return `None`, when the general algorithm can compute the actual Some(_).

Also, checked arithmetic was used when it is not needed, since the
initial guards are enough to assure no overflows will occur.

Tests where changed to include these cases.