Let `in_place_scope` deal with a global pool using the current thread
If `in_place_scope` ends up initializing the global pool because the
current thread is `None`, then we have to recheck the current thread in
case the global is using it -- as may happen on WebAssembly targets.

This fixes the `skip` doctest on WebAssembly, but *any* use of
`in_place_scope` could have had the same problem.